tests/ux/requirements.txt
pywinauto>=0.6.8
mss>=9.0.0
Pillow>=10.0.0
numpy>=1.24.0
pytest>=8.0.0
pytest-timeout>=2.2.0
